Vista articulos

+++ src/main/java/com/hotel/repositorio/ServicioRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hotel.modelo.Servicio;
 
-public interface ServicioRepositorio extends CrudRepository<Servicio, Integer>{
+public interface ServicioRepositorio extends CrudRepository<Servicio, Long> {
 
 }
